backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
python-multipart==0.0.6
python-dotenv==1.0.0